% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_query.R
\name{bbox_query}
\alias{bbox_query}
\title{Create data query for your bounding box}
\usage{
bbox_query(x, years = 2019:2023, columns = "all")
}
\arguments{
\item{x}{\code{sf} / \code{bbox} object (output of \code{\link[=draw_bbox]{draw_bbox()}})}

\item{years}{integer vector of length > 0, since 2019}

\item{columns}{Set of columns to retrieve. Can be "all" (default), "reduced" (n=34), or character vector.}
}
\value{
list of Polars queries to be passed to \code{\link[=download_gedi]{download_gedi()}}
}
\description{
Create data query for your bounding box
}
\examples{
bb = sf::st_bbox(c(xmin=6, ymin=42, xmax=6.1, ymax=42.1), crs=4326)
bbox_query(bb, years=2020)
}
