% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_download.R
\name{download_gedi}
\alias{download_gedi}
\title{Download GEDI data}
\usage{
download_gedi(
  x,
  out.dir = NULL,
  cores = 1,
  progress = T,
  require.confirmation = T,
  timeout = 500
)
}
\arguments{
\item{x}{Output of \verb{[bbox_query()]} or \verb{[tile_query()]}}

\item{out.dir}{Target directory path for download}

\item{cores}{Specify cores > 1 for parallel processing (integer)}

\item{progress}{logical, show a progress bar?}

\item{require.confirmation}{logical, warn and require interactive confirmation for tile downloads with over 10 mio points}

\item{timeout}{Download time out in seconds}
}
\value{
none
}
\description{
Download all GEDI data requested in a bbox_query or tile_query,
created using the respective function.
Data is written to disk in a specified directory and can be opened
using \verb{[arrow::open_dataset(out.dir)] |> [dplyr::collect()]}
}
\examples{
# Construct a query for a small bbox
bb = sf::st_bbox(c(xmin=6, ymin=42, xmax=6.1, ymax=42.1), crs=4326)
bbq = bbox_query(bb, years=2020)

# Define a directory and download
out.dir = file.path(tempdir(), "gedi")
download_gedi(bbq, out.dir = out.dir, timeout = 300)

# Read data into memory and rescale
arrow::open_dataset(out.dir) |>
  head() |>
  dplyr::collect() |>
  rescale_gedi()

# cleanup
unlink(out.dir, recursive = TRUE)
}
