% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_interavtive_map.R
\name{draw_bbox}
\alias{draw_bbox}
\title{Draw a bounding box}
\usage{
draw_bbox()
}
\value{
\code{sf} bbox object with Coordinate Reference System
}
\description{
Open an interactive map and select a rectangular Area of Interest for later
download of point data. The underlying map shows the total number GEDI points
per 5x5 degree tile with map layers for each year.
}
\examples{
\dontrun{
bb = draw_bbox()
print(bb)
sf::st_area(bb)
}
}
