% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_misc.R
\name{make_sf}
\alias{make_sf}
\title{Turn downloaded data into sf object}
\usage{
make_sf(x)
}
\arguments{
\item{x}{table with column "longitude" and "latitude"}
}
\value{
\code{sf} object with CRS WGS84 (EPSG:4326)
}
\description{
Convenience function to create an sf object from a downloaded dataset.
Needs to be loaded into memory beforehand using \code{collect()}
}
