% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_misc.R
\name{rescale_gedi}
\alias{rescale_gedi}
\title{Rescale downloaded data}
\usage{
rescale_gedi(x)
}
\arguments{
\item{x}{in-memory GEDI data table}
}
\value{
rescaled data with all columns provided in \code{x}
}
\description{
GEDI data was transformed from float to integers by applying a scale factor
to save memory on the cloud bucket and speed up retrieval. Apply this function
to scale data back to meaningful (floating point) values.
For an example see \verb{[download_gedi()]}.
}
