% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_misc.R
\name{show_gedi_columns}
\alias{show_gedi_columns}
\title{Display available data columns in the GEDI data}
\usage{
show_gedi_columns()
}
\value{
tibble
}
\description{
Returns a tibble with all variables, their description, unit, scale, data type,
valid range, nodata value, and GEDI source (level 2A or 2B)
}
\examples{
show_gedi_columns()
}
