% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_query.R
\name{tile_query}
\alias{tile_query}
\title{Find tiles intersecting your bbox}
\usage{
tile_query(x, years = 2019:2023)
}
\arguments{
\item{x}{\code{sf} / \code{bbox} object (output of \code{\link[=draw_bbox]{draw_bbox()}})}

\item{years}{integer vector of length > 0, since 2019}
}
\value{
\code{sf} object, subset of tile lookup table with tiles of interest
}
\description{
Intersect your bbox with a lookup table for 5x5 degree tiles, which can be used
for download.
}
\examples{
bb = sf::st_bbox(c(xmin=6, ymin=42, xmax=8, ymax=44), crs=4326)
tile_query(bb, years=2020:2022)
}
